+++ server/src/models/recipes/getRecipes.ts
@@ -17,7 +17,7 @@ const getRecipes = async (): Promise<Recipe[] | false> => {
             position: 'asc',
           },
         },
-        RecipeIngredients: {
+        RecipeIngredient: {
           include: {
             ingredient: true,
           },
@@ -27,6 +27,8 @@ const getRecipes = async (): Promise<Recipe[] | false> => {
         },
       },
     })
+
+    console.log(recipes)
 
     if (!recipes.length) throw new NoContentError(ERROR_MESSAGES.NO_RECIPES_FOUND)
     return formatRecipeImages(recipes)