+++ CHANGELOG.md
@@ -15,12 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ingredient type can now be changed
 - Serverside route error logging
 - Fixed image upload to Cloudinary
+- Recipe can be published via edit recipe page
 
 ### Changed
 
 - Removed body-parser package as Express already has these features built-in
 - Archived fastify-server, removed all .env vars for security reasons - improved .gitignore
 - Now 10 MB attachments may be sent to the server
+- Corrected server entity `RecipeIngredients` to `RecipeIngredient`
 
 ### Fixed